spring-rest-2ts: jax-rs implementation
