+++ pom.xml
@@ -48,6 +48,7 @@
         <module>spring-rest2ts-generator</module>
         <module>spring-rest2ts-jackson</module>
         <module>spring-rest2ts-spring</module>
+        <module>spring-rest2ts-jax-rs</module>
         <module>spring-rest2ts-spring-data</module>
     </modules>
     <profiles>
