+++ spring-rest2ts-spring/src/main/java/com/blueveery/springrest2ts/spring/PathVariableEntity.java
@@ -1,0 +1,14 @@
+package com.blueveery.springrest2ts.spring;
+
+import org.springframework.web.bind.annotation.PathVariable;
+
+import java.lang.annotation.Annotation;
+
+public class PathVariableEntity extends MethodParameterEntity implements PathVariable {
+
+
+    @Override
+    public Class<? extends Annotation> annotationType() {
+        return PathVariable.class;
+    }
+}
