+++ spring-rest2ts-spring/src/main/java/com/blueveery/springrest2ts/spring/RequestBodyEntity.java
@@ -1,0 +1,23 @@
+package com.blueveery.springrest2ts.spring;
+
+import org.springframework.web.bind.annotation.RequestBody;
+
+import java.lang.annotation.Annotation;
+
+public class RequestBodyEntity implements RequestBody {
+    private boolean required;
+
+    @Override
+    public boolean required() {
+        return required;
+    }
+
+    public void setRequired(boolean required) {
+        this.required = required;
+    }
+
+    @Override
+    public Class<? extends Annotation> annotationType() {
+        return RequestBody.class;
+    }
+}
